MIPS_Logic_Instructions.txt
Not all of these have been checked yet, but they have an exhaustive list of each and every logic condition. They operate in turn, so later instructions only depend on the code of previous instructions so if all the previous tests are successful, it is definitely the instruction being tested that is incorrect.